# perf(compiler): use moduleResolutionCache for Language Service's resolveModuleNames