<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

## Summary

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
#1784 introduced the usage `resolveModuleNames` API from `LanguageService`. I noticed that `resolveModuleNames` API can have a cache called `ModuleResolutionCache` so I went ahead and tried out. The result is amazing:

- Before #1784 and after #1784: the performance is more less similar, sometimes #1784 makes performance even worse.
- With this PR: the performance is even better than before #1784 and after #1784 

**Some figures**

Testing against my company's project which has 2k+ test files as well as the repo source has 2k+ files.

After #1784 

```
 PASS  src/__tests__/pulse-items/pulse-items-filter.helper.spec.ts (5.736s)
 
 PASS  src/__tests__/hybrid-apps/widget-card-settings.handler.spec.ts (21.57s)
 
 PASS  src/__tests__/pulse-items/pulse-items-ws-res.handler.spec.ts (31.41s)
 
 PASS  src/__tests__/inbox/inbox-detail-traveller.effects.spec.ts (31.892s)
 
 PASS  src/__tests__/task-viewer/task-viewer.effects.spec.ts (31.972s)
 
 PASS  src/__tests__/shared/numeric-field.component.spec.ts (22.818s)

 PASS  src/integration-tests/inbox-message-view/email-custom-action.spec.ts (18.773s)
 
 PASS  src/integration-tests/inbox-message-view/email-standard-action.spec.ts (15.707s)
```

With this PR

```
 PASS  src/__tests__/pulse-items/pulse-items-filter.helper.spec.ts (5.387s)

 PASS  src/__tests__/hybrid-apps/widget-card-settings.handler.spec.ts (12.154s)
 
 PASS  src/__tests__/pulse-items/pulse-items-ws-res.handler.spec.ts (19.322s)
 
 PASS  src/__tests__/inbox/inbox-detail-traveller.effects.spec.ts (20.206s)
 
 PASS  src/__tests__/task-viewer/task-viewer.effects.spec.ts (19.916s)
 
 PASS  src/__tests__/shared/numeric-field.component.spec.ts (20.487s)
 
 PASS  src/integration-tests/inbox-message-view/email-custom-action.spec.ts (13.759s)
 
 PASS  src/integration-tests/inbox-message-view/email-standard-action.spec.ts (11.163s)
```

Some tests gain huge performance boost and overall, the time to execute the tests for the whole project improved about **20%-30%**. 

The difference in time between no cache run and cache run reduces to **no-cache run = x1.5-2x cache run** comparing to **no-cache run = 2x-2.5x cache run** after #1784

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->
Added unit test for checking the usage of `ts.createModuleResolutionCache`. Green CI

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact and migration plan -->


## Other information

I also correct test for `config-set.ts` which reduced coverage in #1784